requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<2.0.0

# Celery and Redis
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0

# Development - Hot Reload
watchfiles>=0.21.0,<1.0.0

# Configuration
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Logging
structlog>=24.1.0,<25.0.0

# OCR - Google Cloud Vision
google-cloud-vision>=3.5.0,<4.0.0

# PDF processing
PyMuPDF>=1.23.0,<2.0.0
Pillow>=10.2.0,<11.0.0

# ML - Sentence Transformers
sentence-transformers>=2.3.0,<3.0.0
torch>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0

# File handling
python-multipart>=0.0.6,<1.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
httpx>=0.26.0,<1.0.0

# Linting and Code Quality
ruff>=0.4.0,<1.0.0
bandit[toml]>=1.7.0,<2.0.0

# SQLite async support (for testing)
aiosqlite>=0.19.0,<1.0.0
